backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests
python-dotenv
scikit-learn
nltk
beautifulsoup4
